requirements.txt
streamlit
pdfplumber
faiss-cpu
sentence-transformers
llama-cpp-python
torch
